les trajet qui n'existe pas renvoie bien un 404 maintenant

+++ API/database_manager/trajetManager.py
@@ -101,10 +101,9 @@ def search_trajet(data: json_dict) -> tuple[json_list, int]:
 
 
 def edit_trajet(user_info: UserInfo, trajet_id: int, data: json_dict) -> tuple[json_dict, int]:
-    try:
-        trajet = get_trajet(trajet_id)
-    except ValueError:
-        return {}, 400
+    trajet = get_trajet(trajet_id)
+    if trajet is None:
+        return {}, 404
     if (not user_info.isAdmin) and (user_info.user_id != trajet["conducteur"]["idCompte"]):
         return {}, 403
 
@@ -124,6 +123,8 @@ def edit_trajet(user_info: UserInfo, trajet_id: int, data: json_dict) -> tuple[j
 def delete_trajet(user_info: UserInfo, trajet_id: int) -> tuple[json_dict, int]:
 
     trajet = get_trajet(trajet_id)
+    if trajet is None:
+        return {}, 404
     if (not user_info.isAdmin) and (user_info.user_id != trajet["conducteur"]["idCompte"]):
         return {}, 403
     query = """DELETE FROM TRAJET WHERE idTrajet = ?"""
